vectorized model
The original quantum filtering implimentation 50 train images and 30 test images took around 7 minutes and then training the model with the filtered dataset for 30 epochs took around 3 seconds.

The vectorized implimentation takes 1.5 minutes to filter and train a model with 60000 train images and 10000 test images. This roughly estimates to over a 1200 times increase in performance.

This should increase the amount of experiments we can run with various quantum nodes.

Let me know if this needs to be done for the 3x3 square input shape as well.